doc: fix core examples
