+++ examples/core-example/custom.ts
@@ -1,4 +1,5 @@
-import { DisclosureFrame, SDJwtInstance } from '@hopae/sd-jwt';
+import { SDJwtInstance } from '@hopae/sd-jwt-core';
+import { DisclosureFrame } from '@hopae/sd-jwt-type';
 import { createSignerVerifier, digest, generateSalt } from './utils';
 
 (async () => {
@@ -7,14 +8,13 @@ import { createSignerVerifier, digest, generateSalt } from './utils';
   // Create SDJwt instance for use
   const sdjwt = new SDJwtInstance({
     signer,
-    signAlg: 'EdDSA',
     verifier,
+    signAlg: 'EdDSA',
     hasher: digest,
+    hashAlg: 'SHA-256',
     saltGenerator: generateSalt,
-    kbSigner: signer,
-    kbSignAlg: 'EdDSA',
-    kbVerifier: verifier,
   });
+
   // Issuer Define the claims object with the user's information
   const claims = {
     firstname: 'John',
@@ -31,11 +31,15 @@ import { createSignerVerifier, digest, generateSalt } from './utils';
   // Issue a signed JWT credential with the specified claims and disclosures
   // Return a Encoded SD JWT. Issuer send the credential to the holder
   const credential = await sdjwt.issue(claims, disclosureFrame);
-  console.log('encodedSdjwt:', credential);
+  console.log('encodedJwt:', credential);
+
+  // Holder Receive the credential from the issuer and validate it
+  // Return a boolean result
+  const validated = await sdjwt.validate(credential);
+  console.log('validated:', validated);
 
   // You can decode the SD JWT to get the payload and the disclosures
   const sdJwtToken = await sdjwt.decode(credential);
-  console.log(sdJwtToken);
 
   // You can get the keys of the claims from the decoded SD JWT
   const keys = await sdJwtToken.keys(digest);
@@ -53,4 +57,26 @@ import { createSignerVerifier, digest, generateSalt } from './utils';
     claim: JSON.stringify(sdJwtToken.jwt?.payload, null, 2),
     presentableKeys,
   });
+
+  console.log(
+    '================================================================',
+  );
+
+  // Holder Define the presentation frame to specify which claims should be presented
+  // The list of presented claims must be a subset of the disclosed claims
+  // the presentation frame is determined by the verifier or the protocol that was agreed upon between the holder and the verifier
+  const presentationFrame = ['firstname', 'id'];
+
+  // Create a presentation using the issued credential and the presentation frame
+  // return a Encoded SD JWT. Holder send the presentation to the verifier
+  const presentation = await sdjwt.present(credential, presentationFrame);
+  console.log('presentedSDJwt:', presentation);
+
+  // Verifier Define the required claims that need to be verified in the presentation
+  const requiredClaims = ['firstname', 'id'];
+
+  // Verify the presentation using the public key and the required claims
+  // return a boolean result
+  const verified = await sdjwt.verify(presentation, requiredClaims);
+  console.log('verified:', verified);
 })();
