+++ examples/core-example/sdjwtobject.ts
@@ -1,0 +1,57 @@
+import { SDJwtInstance } from '@hopae/sd-jwt-core';
+import { DisclosureFrame } from '@hopae/sd-jwt-type';
+import { createSignerVerifier, digest, generateSalt } from './utils';
+
+(async () => {
+  const { signer, verifier } = createSignerVerifier();
+
+  // Create SDJwt instance for use
+  const sdjwt = new SDJwtInstance({
+    signer,
+    signAlg: 'EdDSA',
+    verifier,
+    hasher: digest,
+    saltGenerator: generateSalt,
+    kbSigner: signer,
+    kbSignAlg: 'EdDSA',
+    kbVerifier: verifier,
+  });
+  // Issuer Define the claims object with the user's information
+  const claims = {
+    firstname: 'John',
+    lastname: 'Doe',
+    ssn: '123-45-6789',
+    id: '1234',
+  };
+
+  // Issuer Define the disclosure frame to specify which claims can be disclosed
+  const disclosureFrame: DisclosureFrame<typeof claims> = {
+    _sd: ['firstname', 'id'],
+  };
+
+  // Issue a signed JWT credential with the specified claims and disclosures
+  // Return a Encoded SD JWT. Issuer send the credential to the holder
+  const credential = await sdjwt.issue(claims, disclosureFrame);
+  console.log('encodedSdjwt:', credential);
+
+  // You can decode the SD JWT to get the payload and the disclosures
+  const sdJwtToken = await sdjwt.decode(credential);
+  console.log(sdJwtToken);
+
+  // You can get the keys of the claims from the decoded SD JWT
+  const keys = await sdJwtToken.keys(digest);
+  console.log({ keys });
+
+  // You can get the claims from the decoded SD JWT
+  const payloads = await sdJwtToken.getClaims(digest);
+
+  // You can get the presentable keys from the decoded SD JWT
+  const presentableKeys = await sdJwtToken.presentableKeys(digest);
+
+  console.log({
+    payloads: JSON.stringify(payloads, null, 2),
+    disclosures: JSON.stringify(sdJwtToken.disclosures, null, 2),
+    claim: JSON.stringify(sdJwtToken.jwt?.payload, null, 2),
+    presentableKeys,
+  });
+})();
