+++ examples/core-example/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "references": [
+    {
+      "path": "../../packages/core"
+    }
+  ]
+}
